Updated purple.styles For Correct Color Balancing
I've updated my original file with an update that better suites the original color scheme I was going for now updated to use material colors as per Google's design guideline thanks to feedback from kanna. Also added in extra information regarding what color affects what, can be helpful as a FAQ section.

Please either rename this to "Deadmano.styles" if it is a user theme or feel free to make it an official color and name it darkpurple.styles as I feel it is too dark for actual purple. Else, you can just call it purple.styles as the only similar purple you have is "Lavender" and this still does fall under purple.